Address comments and some more code moving around
