+++ snuba/query/extensions.py
@@ -7,8 +7,6 @@ from snuba.query.query_processor import (
 )
 from snuba.schemas import (
     get_time_series_extension_properties,
-    PERFORMANCE_EXTENSION_SCHEMA,
-    PROJECT_EXTENSION_SCHEMA,
     Schema
 )
 
@@ -34,6 +32,49 @@ class QueryExtension(ABC):
         return self.__processor
 
 
+PROJECT_EXTENSION_SCHEMA = {
+    'type': 'object',
+    'properties': {
+        'project': {
+            'anyOf': [
+                {'type': 'integer', 'minimum': 1},
+                {
+                    'type': 'array',
+                    'items': {'type': 'integer', 'minimum': 1},
+                    'minItems': 1,
+                },
+            ]
+        },
+    },
+    # Need to select down to the project level for customer isolation and performance
+    'required': ['project'],
+    'additionalProperties': False,
+}
+
+PERFORMANCE_EXTENSION_SCHEMA = {
+    'type': 'object',
+    'properties': {
+        # Never add FINAL to queries, enable sampling
+        'turbo': {
+            'type': 'boolean',
+            'default': False,
+        },
+        # Force queries to hit the first shard replica, ensuring the query
+        # sees data that was written before the query. This burdens the
+        # first replica, so should only be used when absolutely necessary.
+        'consistent': {
+            'type': 'boolean',
+            'default': False,
+        },
+        'debug': {
+            'type': 'boolean',
+            'default': False,
+        },
+    },
+    'additionalProperties': False,
+}
+
+
 class PerformanceExtension(QueryExtension):
     def __init__(self) -> None:
         super().__init__(
@@ -50,7 +91,7 @@ class ProjectExtension(QueryExtension):
         )
 
 
-class TimeseriesExtension(QueryExtension):
+class TimeSeriesExtension(QueryExtension):
     def __init__(
         self,
         default_granularity: int,
