+++ snuba/request_schema.py
@@ -1,16 +1,16 @@
 from __future__ import annotations
 
-import copy
 import itertools
-import jsonschema
 
 from collections import ChainMap
 from dataclasses import dataclass
+from deprecation import deprecated
 from typing import Any, Mapping
 
 from snuba.query.extensions import QueryExtension
 from snuba.query.query import Query
-from snuba.schemas import Schema, GENERIC_QUERY_SCHEMA
+from snuba.query.schema import GENERIC_QUERY_SCHEMA
+from snuba.schemas import Schema, validate_jsonschema
 
 
 @dataclass(frozen=True)
@@ -19,6 +19,9 @@ class Request:
     extensions: Mapping[str, Mapping[str, Any]]
 
     @property
+    @deprecated(
+        details="Do not access the internal query representation "
+        "use the specific accessor methods on the query object instead.")
     def body(self):
         return ChainMap(self.query.get_body(), *self.extensions.values())
 
@@ -91,42 +94,3 @@ class RequestSchema:
 
     def generate_template(self) -> Any:
         return self.__generate_template_impl(self.__composite_schema)
-
-
-def validate_jsonschema(value, schema, set_defaults=True):
-    """
-    Validates a value against the provided schema, returning the validated
-    value if the value conforms to the schema, otherwise raising a
-    ``jsonschema.ValidationError``.
-    """
-    orig = jsonschema.Draft6Validator.VALIDATORS['properties']
-
-    def validate_and_default(validator, properties, instance, schema):
-        for property, subschema in properties.items():
-            if 'default' in subschema:
-                if callable(subschema['default']):
-                    instance.setdefault(property, subschema['default']())
-                else:
-                    instance.setdefault(property, copy.deepcopy(subschema['default']))
-
-        for error in orig(validator, properties, instance, schema):
-            yield error
-
-    # Using schema defaults during validation will cause the input value to be
-    # mutated, so to be on the safe side we create a deep copy of that value to
-    # avoid unwanted side effects for the calling function.
-    if set_defaults:
-        value = copy.deepcopy(value)
-
-    validator_cls = jsonschema.validators.extend(
-        jsonschema.Draft4Validator,
-        {'properties': validate_and_default}
-    ) if set_defaults else jsonschema.Draft6Validator
-
-    validator_cls(
-        schema,
-        types={'array': (list, tuple)},
-        format_checker=jsonschema.FormatChecker()
-    ).validate(value, schema)
-
-    return value
